requirements.txt
websocket-client == 0.54
websocket-server == 0.4